requirements.txt
cryptography>=42.0.0
aiohttp>=3.9.0
pyyaml>=6.0.0
pywinpty>=2.0.0; sys_platform == "win32"
uvloop>=0.19.0; sys_platform != "win32"
